Remove requirement to suppress warning
In order to not waste resources on suppressing an E_WARNING and to accommodate debuggers that may ignore the suppression, we can use `stream_select` to check to see if there is a read available before we attempt to accept the socket.

From the php docs on the read parameter: `The sockets listed in the read array will be watched to see if characters become available for reading (more precisely, to see if a read will not block - in particular, a socket resource is also ready on end-of-file, in which case a socket_read() will return a zero length string).`

The warning about using `===` for comparisons is when checking for an error, in this case, we really only want to continue if we have a truthy value. The function returns `FALSE` for errors OR the (int) `number of socket resources contained in the modified arrays`.