Prevent silencing ExitFailure in ProcessContext->join()
If there was an ExitSuccess, it's expected to be 0. If no failure was present, then the exit code is the only thing to be reported.

But, if there was an ExitFailure, and the process did not cleanly exit, the failure most likely was the cause of the unclean exit and should be reported in any case.